delete contents １階層下のコンポーネントをPageへ
